test/requirements.txt
pytest==7.2.2
PyYAML==6.0
setuptools==63.2.0
datefinder==0.7.3